pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "asr"
version = "0.1.0"
description = "Agent orchestrator for incident investigation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "fastmcp>=2.10",
    "langgraph>=0.6",
    "langgraph-checkpoint-sqlite>=2.0",
    "aiosqlite>=0.19",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-ollama>=0.3",
    "langchain-openai>=0.3",
    "langchain-mcp-adapters>=0.1",
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "streamlit>=1.40",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.32",
    "faiss-cpu>=1.8",
    "langchain-community>=0.3",
    "langchain-postgres>=0.0.12",
    "psycopg[binary]>=3.2",
    "numpy>=1.26",
    "apscheduler>=3.10,<4",
]

[project.optional-dependencies]
# Postgres checkpointer is optional: SQLite-only deploys (and CI) skip
# the wheel and the lazy import in runtime.checkpointer never fires.
# Production Postgres deploys install via ``pip install asr[postgres]``.
postgres = [
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg-pool>=3.2",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=7,<8",
    "httpx>=0.27",
    "ruff>=0.7",
    "pyright>=1.1,<2",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg-pool>=3.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator", "src/runtime", "examples"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/orchestrator --cov-report=term-missing --cov-report=xml"
pythonpath = ["src", "."]

[tool.ruff]
line-length = 100
target-version = "py311"
